requirements.txt
# Core
networkx>=3.2
pydantic>=2.0
httpx>=0.27

# LLM
ollama>=0.4

# TUI
textual>=0.89
rich>=13.0

# API Server
fastapi>=0.115
uvicorn>=0.32

# OSINT
duckduckgo-search>=7.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
google-genai>=1.0
python-dotenv>=1.0
